+++ app/src/main/java/com/konkuk/medicarecall/ui/homedetail/glucoselevel/data/GlucoseRepository.kt
@@ -1,26 +1,7 @@
 package com.konkuk.medicarecall.ui.homedetail.glucoselevel.data
 
 import com.konkuk.medicarecall.ui.homedetail.glucoselevel.model.GlucoseResponseDto
-import retrofit2.HttpException
-import javax.inject.Inject
 
-class GlucoseRepository @Inject constructor(
-    private val glucoseApi: GlucoseApi
-) {
-    suspend fun getGlucoseGraph(
-        elderId: Int,
-        counter: Int,
-        type: String
-    ): Result<GlucoseResponseDto> =
-        runCatching {
-            val response = glucoseApi.getGlucoseGraph(elderId, counter, type)
-            if (response.isSuccessful) {
-                response.body() ?: throw IllegalStateException("Response body is null")
-            } else {
-                val errorBody = response.errorBody()?.string() ?: "Unknown error"
-                throw HttpException(response)
-            }
-        }
-
-
+interface GlucoseRepository {
+    suspend fun getGlucoseGraph(elderId: Int, counter: Int, type: String): Result<GlucoseResponseDto>
 }
